COG-532 reverted back moving TestCognitoServiceFacade to another package, removed back Constants class
